requirements.txt
gradio>=4.0.0
groq>=0.3.0
python-dotenv>=1.0.0
gTTS>=2.3.2
SpeechRecognition>=3.10.0
pydub>=0.25.1
Pillow>=10.0.0
requests>=2.31.0 